farmer: Fix review comments.
- refactored benchmarks
- added more logging
- optimized usage of the PiecePublisher
