+++ crates/subspace-farmer/benches/plotting.rs
@@ -13,9 +13,11 @@ use subspace_core_primitives::crypto::kzg::Kzg;
 use subspace_core_primitives::{
     plot_sector_size, Piece, PublicKey, PIECES_IN_SEGMENT, RECORD_SIZE,
 };
-use subspace_farmer::single_disk_plot::piece_receiver::BenchPieceReceiver;
 use subspace_farmer::single_disk_plot::plotting::plot_sector;
 use subspace_rpc_primitives::FarmerProtocolInfo;
+use utils::BenchPieceReceiver;
+
+mod utils;
 
 // This is helpful for overriding locally for benching different parameters
 const RECORDED_HISTORY_SEGMENT_SIZE: u32 = RECORD_SIZE * PIECES_IN_SEGMENT / 2;
